private functions
created a new class called GC that contains everything that was already in the project.
only the bare essentials are made public (for safety).

things called "gc_etc" are now just "etc" and in the GC class as a static member or a type(def).
gc_register/gc_unregister are now root/unroot because "register" is a C keyword.
make_gc is now GC::make.