bug fixes
fixed a bug where GC::delref() could halt if the reference count fell to zero and the deleter called a GC function.

fixed a bug where ctor/asgn move semantics were still flagged as deleted (will be added later, but for now removed so the copy versions will be used).

added the automatic unrooting system to GC::make() - ensures that GC::ptr instances that are held by value by a type T will not be considered roots.
this means the only GC::ptr instances that are considered roots will be the ones declared as local variables (or just not part of a gc allocation), which is exactly the behaviour i was wanting.

made more stuff private.

added a proof-of-concept for what having GC::ptr inside a type would require.