thread safety
everything's implemented and working just fine on one thread, but there are some problems in the thread safety department.
this update fixes several of them (as well as increased GC::make performance), but there are still plenty.

the test file now contains a stupid but very effective test of thread safety where one thread loops allocating and using an object and the other loops calls to GC::collect().