yet another thread safety patch
found another dangerous area of the ptr class - the reset function.
added a non-threadsafe version of delref (__delref).
with this, the entire reset() function can be done under one mutex lock, which is both faster and safer.
i think the nullptr exception we're gettting occasionally in _mark_sweep() is from the delref-xchg-addref not being a single atomic step - fixed.

the problem still persists, but seems to be substantially less frequent (test prog runs consistently for 1-2 minutes at full blast).