bug fix + more safety
the last commit introduced a bug where if the dtor of T in ptr<T> calls a gc function it would halt (mutex was still locked).
this is because the __delref() code was copied from delref(), which uses the GC::mutex object differently to ensure it's safe.

new algorithm makes __delref() add the handle to GC::del_list.
after one or more calls to __delref() (which is async) you must make a (sync) call to GC::handle_del_list() to perform the actual deletions.
it'll handle all the internal logic for making this a safe action.

identified two more areas of thread safety - the ptr copy ctor and dtor.
like the safety issue in the last commit, these were also problematic because they weren't carried out as a single atomic step.
this is now fixed thanks to the new __delref() and handle_del_list() code.