threading bug finally fixed
finally found the source of the threading bug:
when the mark-sweep algorithm runs, it selects everything it's going to delete and adds it to a set.
afterwards, it goes through each item, calls its deleter, and deletes the handle.
2 problems: 1, this leaves dangling pointers in the subset of "marked" "alive" gc pointers, 2, the deleter may need to refer to a handle that's scheduled for deletion.
fix is to go through "marked" "alive" pointers and null any references to objects in the delete set and to delete the handles after calling all their deleter functions in turn.