minor touch-ups
added bad alloc safety to GC::__create() via std::unique_ptr.

changed the == and != overloads for ptr<T> to directly use handle instead of needing to call ptr<T>::get().
unique objs will always have unique allocation entries and vice versa.
also, == and != are the only comparison operators that are likely to be called, so they should be optimized as much as possible.