auto collect bug fix
in the last commit, the auto collect thread was created as a static object to ensure it was created at prog start without needing the user to do anything.
however, being owned by value means it would eventually call the destructor, which calls std::terminate() if the thread is still alive (which it always would be).

new code has the thread owned as a dynamic object and held as a local static object of a function __start_timed_collect() that begins the process.
this function is called by all invocations of GC::make(), which means the auto collector thread won't be created until the first invocation of make() (so no performance losses whatsoever on code that doesn't use the gc library but for some reason includes it).