security patch + bug fix
changed router_fn to a struct instead of a typedef and hid all of its internal stuff.
because of this, the type hackery in test.cpp only works from a private member of GC (e.g. a router specialization).
but it's already labeled as und to do anything other than calls to GC::route() from a router specialization, so it's ok.

fixed a bug where we were using = for initialization of atomic types (g++ was complaining).
should be legal, but g++ was converting the init value into an atomic via conversion ctor and then reaching the deleted copy ctor.