hotfix
fixed a stupid error on my part where i allowed any static cast conversion but didn't actually account for it (it's slightly subtler than that, but not by much).
the old GC::ptr<T> stored only a GC::info*, which has a void* to the object.
the old method for getting the object from a ptr<T> was to use handle->obj, however many static conversions will have a different address, so this isn't safe.
in particular, multi-inheritance and polymorphic types would be a problem.

new code to fix this: just store a T* inside each ptr<T> as well as the info*.
this has the effects of fixing the problem and speeding up accesses by probably around 2x due to having half the indirection per access (not tested).
cost is ptr<T> takes twice as much space as before and needs a bit more care in management, but that's really unimportant.