perf + prep for allocfail strat
removed __create() (which required a lock and allocated the info object before linking it into the gc database).
the alloc is now done in make() and is non-blocking.
the linking code still requires a lock and is now in its own function __link().

not having an allocation in the lock seems to be a moderate speedup, though this isn't the primary reason for the change:
added the strategies::allocfail flag that performs a GC::collect() any time GC::make() has an alloc failure.
this isn't currently implemented - needs one more tweak to the allocation code before it's "really" doable in an efficient manner.