added support for ptr<T[N]>
added support for ptr<T> where T is a C-style array
added several utilities for dealing with this:

construction/destruction is now handled by a new type called __ctor_dtor<T>.
it has static methods ctor() and dtor() that take pointers to location to construct the object (in place).
the scalar version ctor takes extra args and passes it to a C++14 unifor initializer syntax placement new expression.
the array version ctor calls the scalar version for each position in the array.
dtor functions for both are much the same, but take no args, of course (array dtor calls scalar dtor).
the reason placement new can't be used for arrays is because (while legal in visual studio) it's not standard C++ and is und or even outright illegal on other systems.
being templates, these ctor/dtor utilities correctly account for any level of multi-dimensional array.

added __full_extent<T> as a std::integral_constant<std::size_t> to get the full extent of a multidimensional array.

added a specialization of router<T> for when T is a C-style array - just calls route() for each element in the array.

added a test case for populating a ptr<ptr<int>[2][2][2]>.