added support for const types in route()
support was added for allocating const types in one of the last updates, but router() wasn't updated.
this means, given a const object, it would use the default specialization of router<T>::route() that is no-op.
this, of course, is terrible.

the new update changes a bit of everything to enable anything to be const.
this includes changing router_fn to take info*const& and roots to contains info*const*, among other things.
the T in router<T> is still non-const, but obj in router<T>::route(obj, func) must be const T&.
GC::route<T>() now takes const T& and forwards it to router<T>::route() where T has its constness removed.
this still preserves const-correctness because the obj that is passed is itself still const (but passed to a handler that is named by non-const type T).