unrooting bug fix
how the rooting system works is: creating a ptr<T> marks it as a root, but the allocation by GC::make<T>() will go in and unroot anything it can route to via router<T>::route(), which makes a ptr<ThingThatHasAPtr> only register as 1 root.
this is required because roots can never be deallocated by mark-and-sweep, and T might contain a ptr that points to its parent, which would be an edge-case irreducible cycle.

however, this is only done for construction by a call to GC::make<T>() - if a ptr is added to the object after construction, it will still be erroneously marked as a root.
this can happen because e.g. T contains a std::vector<GC::ptr<J>> and an item is added after construction.

to solve this, GC::collect() now has an additional step prior to the mark-and-sweep: claim all objects' children.
it does this by performing the same unlinking operation GC::make<T>() does for each element in the gc database (inside the mark clear pass).

realized the unroot lambda had the same signative as __unroot, so it's used directly instead (short and sweet).