added alignment-depended allocators
the old method of allocating space for objects in GC::make<T>() variants was to always use a force-aligned allocator (i.e. GC::aligned_malloc).
this has the downside of taking up at least a pointer's worth of space in the best case scenario for no good reason on each and every allocation.
additionally, the vast majority of types used in C++ are standard alignment, which is already guaranteed by dynamic alloc functions.

a type called __passive_aligned_allocator was added that wraps std::malloc and std::free.
a type called __active_aligned_allocator<std::size_t align> was added that wraps GC::aligned_malloc and GC::aligned_free.

a type called __checked_allocator<typename allocator_t> wraps another allocator in cpp-gc's specific allocation fail-case logic.

a function called __returns_checked_aligned_allocator_type<std::size_t align>() takes an alignment and uses sfinae to select which allocator to use, and additionally wraps that in a __checked_allocator.
these functions have no definitions and are only meant to be used in decltype contexts.

with all of these things defined, we just need to use the return type of the sfinae overloads to get the most efficient allocator to use, which already has all the cpp-gc fail case logic built-in.

this, along with the last vtable update is the end of the obvious space saving tactics.
put together, we should be saving around 3-4 pointers' worth of data on each allocation (plus a (negligible) speed increase for the simplified allocators).