changed info_vtable funcs to take info&
changed the function pointers in the info vtable to take the info object itself instead of the object.
this is a necessary change for the upcoming GC::adopt() feature that will gc manage a pre-existing object (e.g. from a new expression).
that function will need to allocate an info object separate from the object, and there will be no way to get from object to info, but info contains a pointer to the object so that direction is trivially doable.