concept phase for GC::adopt
added the first impl for GC::adopt() - i'm already seeing some problems with usage safety, but it shouldn't be too difficult to get it into a suitable form.

it'll need an overload for taking dynamic arrays, but there's no way to type-determine dynamic arrays, as they're identical to pointer-to-first-element in every conceivable way.
as near as i can tell, it'll be up to the user to specify if it's an array or not, and if so whatever the size is.
however, this presents a ton of safety issues ...