bug fix + new array usage
fixed a stupid bug on my part where the ptr<T> cast functions (e.g. reinterpretCast) weren't incrementing the reference count.
this was because they use ptr<T>::__init() for efficiency, which is only designed for use directly by the make() function family.
this is solved by adding explicit __addref() calls in the cast functions.
a warning was added to the __init docs to emphasize this risk.

this bug was encountered while trying to make the bound array form of the make() function, which for simplicity is defined in terms of the unbound array and just reinterpret casted to the correct type.
a test is added for this case (the above bug) - failing said test is a runtime error.

added array bound helpers to remove bound/unbound extent specifically (doesn't touch the other).
GC::ptr<T>::element_type is now defined as GC::remove_unbound_extent_t<T>.
this means ptr<T> holds T*, ptr<T[N]> holds T(*)[N], and ptr<T[]> holds T*.
this (slightly) goes against the standard library smart pointers, but i believe it more-closely resembles raw pointers, which is the intent.
from here on out, ptr<T[]> is considered the array form and all other constructions are scalar.
i.e. ptr<int[6][5]> is a pointer to a single int[6][5], not an array of several int[5].