added atomic specialization
added a specialized class for atomic<GC::ptr<T>>.
has all the bare stuff with mutexes and lock_guards - pretty trivial stuff.
left out the memory sequencing stuff since it's not relevant for our purposes (some functions take it to be standard-conforming but ignore it).

added a test case for this in the test file - allocates many 64K int arrays over and over again from two threads and stores to a global atomic instance.