added router functions for atomic types
added an ill-formed catch-all specialization for std::atomic<T> because under normal circumstances to route to it we would have to load the current contents of the atomic type; however if the load function calls any gc functions it could deadlock.
thus - in general - routing to atomic types is forbidden.

but we of course need ours (i.e. std::atomic<GC::ptr<T>>) to work.
fortunately, we wrote this type and can give ourselves friend status to access the internals.
thus, we lock the internal mutex and route to the stored object directly, thus bypassing the gc function calls from the load function.