atomic deadlock bug fix
found a bug in the new atomic code that can cause deadlocking.
it was because of having two mutexes that needed to be locked in different order for their independent logic to work.
the fix rewrites the atomic code in such a way that it doesn't have an extra mutex (probably just as efficient due to having to lock the GC::mutex anyway).

however, this had to be done as a gc internal class:
added GC::atomic_ptr<T> to do this - does the actual atomic logic.
std::atomic<GC::ptr<T>> is now a wrapper for GC::atomic_ptr<T>.

re-ran the test case that caused the potential deadlock and couldn't repeat the error after this patch.