beginning bug fix branch
this commit adds a test case that (quickly) results in a deadlock.
this is due to the nature of how i go about supporting thread safety (which is clearly not very good).
looking for a solution.

the easiest solution would be to open up the low-level interface to the users, but that would be dangerous, stupid, and annoying.
so, i'm now looking for a way to make everything work or its own independent of anything the user could possibly do - let the fun begin.