major thread structure revamp
rewrote all the thread safety logic to use different locks for each database object.
this is controlled via new (private) database types that encapsulate the bahavior and each have their own mutex.

this doesn't fix the issue quite yet, but it's a big step on the way.
for instance, this commit makes it a lot clearer which parts of the gc system require others and which are standalones.
e.g. apparently the object database is the only thing that needs to be mutex from mutable router insertions.
this will be exploited soon to fix the problem.