fixed the deadlocking router bug
fixed the deadlocking router bug, but needs a ton of testing.
this required a change over to using std::atomic<info*> instead of raw info* for the ptr and roots database classes.

in particular, testing needs to be done to make sure repointing a GC::ptr during garbage collection is safe.

on the bright side, if all of this is safe it could potentially mean a lot less locking (or at least a lot less locking of the same lock from multiple sources).

now i'm suspicious of atomic_ptr, which does its own locking on the objs database.
this'll likely need to be changed to use a different mutex to perform the internal synchronization.