added the missing handle_del_list call
added in the missing handle_del_list() call in delref().
after reviewing it, i don't think there's a problem with doing it directly inside delref().
the only case i can see that would be a problem is if it's called during a router function, but i've already flagged doing that as undefined behavior, so we're good.