another deadlock bug
changed the test file to a new configuration that results in a deadlock (really quickly).
this one's caused by the ptr funcs calling delref, which calls del_list.handle_all(), which locks the objs_database mutex.
this wouldn't ordinarily be a problem, but if the host object locks its own mutex(es) in the router and the mutator, this can result in locking them and the objs mutex needed for the router functions in different orders, which causes a deadlock.