more bug stuff
added a bunch of stuff to try to fix the other bug, but meanwhile found a new bug that might be causing some of the crashes in attempting to fix the first.

this one involves needing the roots db be locked to perform collection logic, but the other thread might lock a mutable router mutex and try to add/remove a ptr, which would require a root db hit.
there's not a way to make that any better, so i just need to rewrite the roots db so that this isn't a problem.