roots db fix
changed the roots db - now uses separate sets to represent queued add/remove operations.
the only thing that uses the roots database contents is collect(), which still takes a mutex for safety reasons.
the queue sets use a different mutex that is marked as "do not lock this ever" for external sources.

when an add/remove op is made, it goes into the queue sets.
when the collector needs the real contents it calls a helper function to merge the sets together and effectively perform all the queued ops.

this fixes the issue where it was deadlocking really quickly, but apparently there's now (or there already was) an error that causes a crash after about 40 seconds in debug mode - looks like a dead object access.