added smart_handle class
had a horrible realization - made an example diagram using the current system that showed it could easily result in premature deallocations.
now pretty much everything except the make/adopt functions and the router/info system needs to be scrapped and reworked.

added the smart_handle class (not fully implemented).
the only part not implemented is the reset function, which needs to somehow be simultaneously non-blocking and mutually exclusive with any collector thread.