new impl for everything - mostly working
added the new implementation for everything - it's now all mostly-working.
in debug mode it runs fine, but there's an occassional exception in the release build that'll need investigating.

as was stated in the previous commit, we need a method of repointing handles that is mutex with the collector to keep it from messing up its internal book-keeping mid-way through a graph traversal, but we also can't use traditional locking algorithms because router functions are allowed to lock arbitrary mutexes, which could potentially cause a deadlock.
thus we needed a new algorithm that allows non-blocking handle repointing but that also won't modify it during a collection pass (but still works during such).

this new logic is now in the code base, but continuous changes are a bit annoying, so the controls for this feature are rolled into a new wrapper class called smart_handle that controls all of that internal logic.

as for the algorithm, a handle exists somewhere (e.g. inside a ptr<T>) and points to the object that should be used for reachability traversals.
all handle-repoint requests are sent to a separate cache instead of applied directly to the handle - internal logic ensures that the cache reflects the net change for each pointer, including the case where the handle is destroyed.
when the call is made to begin a collect cycle, the net changes in the cache are applied to the actual handles and the cache is voided.
this all happens inside the collection_synchronization class (renamed from collection_deadlock_protector), which ensures that these actions can't conflict with one another.