added roots to sync helper
added the root/unroot system to the collect_synchronizer class and optimized it (there's no reason anyone would need to re-root and unrooted handle).

there's still a crash that persists, which is now in the debug version.
honestly it's reasuring that it still happens in debug mode because there're more tools to find it that way.
once that bug's fixed (assuming nothing else goes wrong), this bug fix branch should finally be done.

the begin_collection() and end_collection() functions were a bit too easy to mess up and don't offer much in the ways of encapsulation.
new sentry subtype called collection_sentry does all of this work automatically - you just need to convert it to bool to makes sure it's good.