stupid bug fix
fixed a bug in the new root system that resulted from copy/paste errors (the create() function was a duplicate of the destroy() function and wasn't rooting anything).
verified that everything is working in one thread, so as long as the syncronizer class does its job it should be fine for the next run.