more bug fixes
found a sneaky bug that was caused by the smart_handle destructor being non-blocking.
it has to be this way for everything to work, but it also means its contents can become undefined.
fixed this by making the info* handle a dynamic allocation and wrap the behavior around info** to work the same way.
the schedule_handle_destroy() function now adds the address of the raw handle reference to a list that's deleted by the collection sentry constructor.