yet another bug fix - who's next!?
fixed the last bug that was crashing the (stripped down) test file.

the bug was caused by new objects being inserted into the obj database before their handles were rooted and that step not being atomic with respect to the start of a collection action from another thread, so the object would seemingly have no roots (roots would be in the cache because they came in too late) and would thus be immediately deleted.
then, when the roots were injected for the next pass (assuming the deletion phase's handle dtor was also too late), the info object would be deleted and the mark sweep would access undefined memory.

still a crash problem - of course -_-