fixed a couple of problems - still more
fixed a couple of issues - one big one was the use of smart_handle.raw_handle() to get a smart_handle ctor argument (very wrong).

this is fixed by a new syncronizer function schedule_handle_create_alias(new, src) that atomically creates 'new' as a new handle (rooted) and immediately repoints the actual pointer to src's current target - this is safe because 'new' is a new raw handle and is thus not already in the gc database anywhere - thus since we're atomically adding it we can safely repoint it immediately.

moved the collect delete logic into the collection sentry dtor - this means a del_list class is not needed and if the dtors for the objects trigger a collect action it'll be no-op (good thing).