fixed?? merged obj database into collect sync
renamed the obj database type to obj_list and merged it into the synchronizer class.
it now has the same logic as the other sync stuff - where there's a collect runtime list of objects and an obj add cache, which potentially speeds up allocations.
additionally, this means no operation is blocking, so you can gc allocate new objects while the collector is still running without any repercussions.

i actually only did this as a potential fix for the bug that i can't seem to find.
if this doesn't fix it i'm totally stumped...
on the bright side, it seems to be working (kinda hard to tell for certain since it's a concurrency bug), but i still don't understand why the old code didn't work.

removed some stuff that could be dangerous, which might have been what was causing the problem, but i couldn't find any occurrences of it that weren't guarded to the point that the dangerous aspect could never happen.

the biggest aspect that i think fixed it was making the method of adding a new object atomic with respect to its handle root operation, but again, the object insertion was always after the root operation, so i don't see how that could possibly not work...

whatever