ughhhhhhhhhhhhhh
yet another crash bug that doesn't seem like it's even possible -_-

this one is apparently from trying to remove something from the collection obj list that isn't in the list.
however, the collection obj list is heavily guarded and i can't find any accesses to it that could even remotely be considered dangerous.

there's also another bug where visual studio says it's detecting stack buffer overflow, but i can't even imagine how that would happen since literally nothing we're doing is on the stack... this is potentially (and hopefully) the same bug as above).

on the bright side, this bug is really uncommon and i have to run the program several times to get it to happen... eh...