nvm - i'm retarded
apparently the bug from the last commit was in the merge code (new function that hadn't been tested).
it was so simply i didn't feel the need to actually test it.
apparently i'd forgotten to repoint the "last" pointer for the destination of the merge in the case where the target is non-empty.

having fixed that, it seems to work perfectly - all other test cases have been un-commented and seem to work flawlessly.
i'm not sure what was going on with the only other error (stack corruption) but i haven't been able to reproduce it since making this change, so that was probably the only problem.

so looks like everything's fine to merge back into the master branch. yay.