more stuff + bug fixes
found a few places in the unique_ptr wrapper where we were breaking thread safety + one in atomic_ptr where we were breaking atomicity.
these were both due to only locking the "this" mutex when we should also lock the other's as well if we're mutating it.

the easy fix is to use std::scoped_lock to lock both mutexes and ensure no inversion deadlocks, but cpp-gc is made for C++ versions prior to 17, so we can't do that.
so i made my own implementation of it GC::scoped_lock (public) that functions identically to the std version.
this doesn't have any weird gc logic - it's just in the gc class to prevent name collisions (i imagine it's quite common for code bases to define their own implementations of this class for pre 17 builds).

GC::scoped_lock was then used to fix these errors as well as replace a few direct use of std::lock() that were only added because std::scoped_lock was unavailable.

added GC::vector as a wrapper for std::vector - not nearly finished - still has the same bugs the old GC::unique_ptr had in terms of thread safety.
this commit started with adding that, but then changed to be bug fixes - vector will be finished in the next commit.