more wrappers + "bug" fixes
added the deque wrapper.

fixed a "bug" where wrappers were acting slightly differently from the object they wrapped (e.g. a constructor might actually be implemented using assignment).
this is now fixed and should be identical in every way to using the underlying type (with the exception of the occasional mutex lock).
this is facilitated by using a buffer to hold the object and constructing it explicitly with placement new after the mutex is locked (otherwise we would have to default-construct the object and assign it after the lock like we were doing before).

added a few minor test cases for some of the wrapper types - they'll get more extensive testing later on.