finished stdlib wrappers
added unordered_map and unordered_multimap wrappers - this finishes the full set of wrappers.

now they just need to be reviewed and tested in various versions of C++ (e.g. .contains() is a C++20 method in e.g. std::set, but visual studio has it available in C++14 mode as well).
mainly, if it runs in the linux build it's good to go.

perhaps the biggest issue is that methods of a template class aren't compiled unless they're used, which means unless i make test cases for using literally every single constructor, operator, function, and conversion for every single wrapper, there may be occassional compile errors down the road that i'll need to hotfix (very easy to fix, but extremely tedious to test for exhaustively).

the next few commits will concern efficiency and bug fixes in the wrappers.