optimization
halfed the size of GC::scoped_lock by using a different unlocking method.

the problem is you need to keep the type information as well as the unlock arguments, which i was doing with a pair of a function pointer taking a void* and a void* for the argument - the function pointer encapsulated the type.
the new method uses std::tuple<T&...> to hold the arguments and keep their type info.

i had thought about doing this the first time around, but i thought the overhead of having a bunch of recursive calls in the unlocker would bog down the efficiency and effectively trade time for space.
however, after analysis via godbolt, this doesn't appear to be the case, and the compiler can easily optimize it away even for 8 mutexes (which while not a lot is more than you'd typically have need for).
and even if it can't optimize it all away, it uses tail recursion, so we'd be using jmp instructions instead of call, thus making the spatial complexity O(1).

added struct __scoped_lock_unlocker to assist with the unlocking process.