wrapper typo fixes
fixed a few typos in the wrappers.
the only "big" mistakes were missing "return" keywords for the GC::list insert() return values.
upon use, this would likely raise warnings (but not errors - at least in vs18), so it's good this was caught early.
only other mistakes consisted of capitalization errors on type names in the disabled preprocessor blocks for the C++17 stdlib wrapper classes.

all the wrappers have now been double-checked for mutex safety.
this'll need ongoing work to keep up-to-date with the changing standard (as well as add backwards compatibility for older versions, though we only support C++11 at the VERY least).

for the forward_list and list wrappers, added merge/splice functions that take a wrapped_t object instead of only the gc version.