more bug fixes
fixed all the new stuff to work in the linux build.
biggest change is that i had to move (and rename) the scoped_lock custom code - now above the gc class - this is because i forgot gcc (correctly) doesn't allow full class specialization except at the namespace scope.
however, an alias was added for it in the gc class, so no other code needs to be updated and we can pretend it never happened.

fixed a few cases (all together) of using the wrong type name for the stdlib wrapper comparison operators.