bug fix
fixed a known bug that's been in the collect logic since the massive sync class rewrite - here's the problem:
the collector was using GC::router_unroot(), which uses the schedule_handle_unroot() function in the sync class.
however, this means that the collector was only scheduling unroot operations, instead of doing them instantly.
this doesn't present any massive problems, but it does mean it can take up to 2 passes to correctly identify an object as garbage and delete it.
added a test case to demonstrate this.

the fix was to add a new collect sync method called unsafe_immediate_handle_unroot(), which directly modifies the collector-only data (only safe to use if the sentry has successfully been constructed because we must have exclusive access to the collector-only data, obviously).
additionally, because it already assumes we have exclusive access to the collector-only data, it doesn't require a mutex lock.

all of this together fixes the problem and drops cpu usage of the test program from 33% to 25% and with higher throughput (vsc++ mutexes are quite inefficient).