optimization
added a memory usage (and probably time to a lesser extent) optimization to the handle repoint and root/unroot sync functions.
the logic is: if there's not a collection action in progress, we don't have to cache the request, we can just do it immediately.

this works for the root/unroot and repoint actions, which this commit updates.
this also applies to the obj list insertion, but since we use obj_list::merge(), there's really no need to bother changing that - it would just add extra code branches and wouldn't actually save any memory.

i believe this logic would also work for the handle deletion cache, but i'll need to double check that - it's not applied in this commit.