bug fix from last commit + more optimization
fixed a bug caused by the last commit.
the problem was that the cache logic was a bit too weak.
this meant that we might perform immediate changes even though there were scheduled changes in the cache from previous invocations.
the result of this would thus be incorrect ordering (i.e. a = b; a = c; could result in a net change a = b, which is of course very wrong).

the fix for this was just to strengthen the cache logic.
caches are no longer allowed to be non-empty outside of a collection action.
so, when there's no collect action you MUST perform the change immediately.
upon terminating a collect action, all caches are flushed in the same atomic step.
all of this also supports minimizing memory usage.

however, the biggest change (why this commit took longer than usual) is in handle optimization.
previously, there was a problem with handles: namely we would iterate over them in the sweep logic but they might have already been destroyed (and would thus be dangling pointers).
the solution i used to fix this was to make handles be dynamic resources (i.e. info** instead of info*) - this fixes the problem but results in a ton of dynamic allocations.

another (possibly better) solution that i just thought of while writing this is to lock the internal mutex for each iteration over the roots and only use it if it's not in EITHER of the root scheduler caches.
however, this would result in a ton of locks and would likely be extremely inefficient because of it (though it wouldn't require all the dynamic allocations).

anyway, the solution applied in this commit (which i think is objectively the best) is just to not use the roots set.
we add a new set that contains the actual rooted objects (i.e. objects that rooted handles point to) and use that instead.
this fixes the problem of the handles being destroyed (and thus undefined) and dereferenced by not actually dereferencing them at all.

this new set is called root_objs and is populated in the same atomic step as the cache flushes in the collection sentry ctor.
this avoids both the dynamic handles problem and the mutex locking problem.

theoretically, i can't see any problems with it.
experimentally, this has been thoroughly tested and seems to work perfectly for the test file.
the last commit caused an undefined memory access error ~300ms into debug execution (without the cin.get()) on approximately 1/20 executions - this optimization/patch was executed in both debug and release modes around 80 times each and could not repeat the error.