re-added reference counting logic
about 40-50 commits ago there was a catastrophic series of bugs that required virtually all of the memory management code to be scrapped and rewritten.
among the things removed was the old reference counting system for object deletion.

under the od system, the reference count for any object falling to zero would trigger the immediate deletion of the object.
but this caused problems with mutexing and dangiling pointers etc. which is why it was left out.
now, however, i've re-examined some logic and decided it would be beneficial and easy to re-implement now that everything else is already working bug-free.

the new system will manage the reference counts and - when it hits zero - will delete the object in a blocking manner on its own if there's not currently a collection action in progress (i.e. GC::collect()).
however, if there's a collection action in progress, we can't delete it in a blocking manner because the garbage collector is using those resources.
thus, if a collection action is running, the object isn't deleted.
but rest assured: having a reference count of zero necessarily implies it's unreachable, so the garbage collector will dispose of it.

the test case now runs perfectly in debug and release mode and uses constant memory without needing to call collect() at all (in the long-running part at the end).

i'll do a bit more work on it later and see if i can make the deletion blocking regardless of if a collection action is in progress.
i have a hunch that it may be doable with a mutex on the obj list, but that could potentially hurt collector performance with N locks where N is the number of gc objects.
also that might require a different data structure for obj_list (probably a set or hash set) because we wouldn't know in that case if the object is in the obj list or the obj add cache.

i also need to test an edge case of the reference count logic: can the reference count decrementer not delete something because a collection action is running and said collection action also not delete the object by the end.
if this is the case i'll need to add yet another cache to hold pointers to the info objects it planned on deleting and have the collector deleter handle those as well.