fixed a ref count bug
found and fixed a bug with the new ref count system that was speculated about in the last commit.
the old system would delete the object immediately if there wasn't a collection action, or would do nothing otherwise (thus leaving it up to the collector thread to dispose of it).
the bug was that, depending on what things happen when, it could leave it up to the collector but the collector could (in its now-old snapshot of the object graph) determine that the object is not garbage.
thus, the reference count fell to zero but the object was never deleted.

of course, if calls to collect() are infrequent, this isn't a problem because it can only occur during a collection action.
also, if calls to collect() are frequent, this isn't a problem because the next call to collect() will correctly flag it as garbage.
however, in trying to create the best product possible and seeing how everything else was already working, this was fixed with a few data structure changes and some brand new logic.

the solution was to have yet another cache for specifying objects that should be ref count deleted that fall into the category in which the bug occurs - which to be absolutely specific is during a collection action for which the object is not in the obj add cache and thus under active gc consideration.
however, we need to manage this cache in strange ways to make sure everything happens when it's supposed to.
one possible solution would be to lock, get the cache contents, unlock, and handle it.
however, in that time the cache might have more contents, so this would have to be done in a loop that can only terminate when the cache is actually empty.
this is, of course, unacceptable.

the better solution (which is in this commit) is to have special caching logic with a different flag determining if we should cache or apply ref count deletions directly and immediately.
this is similar to the old system of "are we in a collection action" except that its duration is a subset of the collection action duration.
thus, when we start a collection action we also start a ref count del caching action.
after performing the graph traversal and destroying the unreachable objects we end the ref count del cache action (thus resuming ref count logic in the old style as if we weren't in a collection action).
we then, in the same atomic step, traverse unreachable objects and purge them from the ref count del cache before performing the final steps of deallocating unreachables and unlinking/destroying/deallocating the ref count del cache contents.

it's honestly not that bad now that i lay it out like this, but it does mean a lot more moving parts for future updates to the low-level interface.